Moved helper function 'proxy_mysql_stmt_close' from 'gen_utils' to 'MySQL_Protocol' #3525
