+++ include/gen_utils.h
@@ -230,22 +230,6 @@ inline unsigned long long realtime_time() {
   return (((unsigned long long) ts.tv_sec) * 1000000) + (ts.tv_nsec / 1000);
 }
 
-/**
- * @brief ProxySQL replacement function for 'mysql_stmt_close'. Closes a
- *   MYSQL_STMT avoiding any blocking commands that are sent by default
- *   'mysql_stmt_close'.
- *
- *   NOTE: This function is not safe, caller must check that the supplied
- *   argument is not NULL.
- *
- * @param mysql_stmt An already initialized 'MYSQL_STMT'. Caller must ensure
- *   that the supplied argument is not NULL.
- *
- * @return The result of calling 'mysql_stmt_close' function over the internally
- *   modified 'MYSQL_STMT'.
- */
-my_bool proxy_mysql_stmt_close(MYSQL_STMT* mysql_stmt);
-
 #endif /* __GEN_FUNCTIONS */
 
 bool Proxy_file_exists(const char *);
